Refactor discovery into own npm package, define common modules outside packages
