+++ README.md
@@ -1,36 +1,18 @@
 # dsatter
 Distributed Systems course project
 
-## Setting up environment
+## Environment setup
 ```
 git clone git@github.com:kordaniel/dsatter.git
-npm install
+cd node-discovery
+- follow instructions in README.md
+cd node-server
+- follow instructions in README.md
 ```
+node discovery service must be running before attempting to run the node-server instances or clients.
 
 ## Nodes
 ### Node discovery service
-Provides an REST api for querying and (un/)registering active nodes statuses. Must be started before the chatserver nodes.
-```
-GET  localhost:8080/api/nodes/active
-POST localhost:8080/api/nodes/active/register
-POST localhost:8080/api/nodes/active/unregister
-```
+Defined in the directory `node-discovery`.
 ### Chatserver nodes
-No error handling implemented. Currently connections are not always terminated properly in both ends. 1..n nodes can be started and they will open a connection between every running node.
-
-## Dev environment
-### Configured commands
-Currently one configured npm project in the directory `node-server`.
-### Dev environment
-```
-npm run lint
-```
-#### Node discovery service
-```
-npm run discovery
-```
-#### Chatserver nodes
-```
-npm run cs            - uses nodemon
-npm run chatserver    - terminates normally
-```
+Defined in the directory `node-server`.
