+++ node-discovery/.eslintrc.js
@@ -1,0 +1,48 @@
+module.exports = {
+  'env': {
+    'commonjs': true,
+    'es2021': true,
+    'node': true
+  },
+  'extends': 'eslint:recommended',
+  'overrides': [
+  ],
+  'parserOptions': {
+    'ecmaVersion': 'latest'
+  },
+  'rules': {
+    'indent': [
+      'error',
+      2,
+      {
+        'SwitchCase': 1
+      }
+    ],
+    'linebreak-style': [
+      'error',
+      'unix'
+    ],
+    'quotes': [
+      'error',
+      'single'
+    ],
+    'semi': [
+      'error',
+      'never'
+    ],
+    'eqeqeq': 'error',
+    'no-trailing-spaces': 'error',
+    'object-curly-spacing': [
+      'error',
+      'always'
+    ],
+    'arrow-spacing': [
+      'error',
+      {
+        'before': true,
+        'after': true
+      }
+    ],
+    'no-console': 1
+  }
+}
