+++ node-discovery/controllers/nodes.js
