+++ node-discovery/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "node-server",
+  "name": "node-discovery",
   "version": "0.0.1",
-  "description": "Node in a network of nodes",
+  "description": "Node serving clients and node-server nodes",
   "main": "index.js",
   "scripts": {
     "start": "cross-env NODE_ENV=production node index.js",
@@ -11,13 +11,12 @@
   },
   "author": "dsatter group",
   "license": "ISC",
-  "dependencies": {
-    "axios": "^1.1.3",
-    "ws": "^8.10.0"
-  },
   "devDependencies": {
     "cross-env": "^7.0.3",
-    "eslint": "^8.26.0",
+    "eslint": "^8.27.0",
     "nodemon": "^2.0.20"
+  },
+  "dependencies": {
+    "express": "^4.18.2"
   }
 }
