+++ node-server/services/discovery.js
@@ -1,5 +1,5 @@
 const config = require('../utils/config')
-const logger = require('../utils/logger')
+const logger = require('../../common/utils/logger')
 
 const axios = require('axios')
