+++ node-server/state/node.js
@@ -1,9 +1,9 @@
 const assert = require('assert')
 const config = require('../utils/config')
-const logger = require('../utils/logger')
+const logger = require('../../common/utils/logger')
 const {
   isNonEmptyArray, sleep
-}            = require('../utils/helpers')
+}            = require('../../common/utils/helpers')
 
 const discoveryService = require('../services/discovery')
